Extract group settings UI
