+++ composeApp/src/commonMain/kotlin/io/music_assistant/client/ui/compose/home/PlayersPager.kt
@@ -19,27 +19,21 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.layout.wrapContentSize
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.pager.HorizontalPager
 import androidx.compose.foundation.pager.PagerState
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.VolumeMute
 import androidx.compose.material.icons.automirrored.filled.VolumeUp
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.Remove
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Button
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.Slider
@@ -67,6 +61,7 @@ import io.music_assistant.client.data.model.client.PlayerData
 import io.music_assistant.client.player.sendspin.SendspinState
 import io.music_assistant.client.ui.compose.common.action.PlayerAction
 import io.music_assistant.client.ui.compose.common.action.QueueAction
+import io.music_assistant.client.ui.compose.home.players.GroupSettings
 import io.music_assistant.client.utils.conditional
 import kotlinx.coroutines.launch
 
@@ -334,180 +329,10 @@ private fun GroupDialog(
         onDismissRequest = onDismiss,
         title = { Text("Group settings") },
         text = {
-            Column(
-                modifier = Modifier.fillMaxWidth()
-            ) {
-                // Non-scrollable Done button at top
-                OutlinedButton(
-                    onClick = onDismiss,
-                    modifier = Modifier.fillMaxWidth().padding(bottom = 8.dp)
-                ) {
-                    Text("Done")
-                }
-
-                // Scrollable list of players
-                LazyColumn(
-                    modifier = Modifier.fillMaxWidth().heightIn(max = 400.dp),
-                    verticalArrangement = Arrangement.spacedBy(8.dp)
-                ) {
-                    // Current player at the very top
-                    item {
-                        GroupPlayerItem(
-                            playerId = item.player.id,
-                            playerName = item.player.name,
-                            isGroup = item.player.isGroup,
-                            volume = if (item.player.isGroup) item.player.groupVolume else item.player.volumeLevel,
-                            isMuted = item.player.volumeMuted.takeIf { item.player.canMute },
-                            simplePlayerAction = simplePlayerAction,
-                        )
-                    }
-
-                    // Bound players
-                    val boundChildren = item.groupChildren.filter { it.isBound }
-                    items(boundChildren, key = { "${it.id}_${it.volume}" }) { child ->
-                        GroupPlayerItem(
-                            playerId = child.id,
-                            playerName = child.name,
-                            volume = child.volume,
-                            isMuted = child.isMuted,
-                            simplePlayerAction = simplePlayerAction,
-                            bindItem = child,
-                        )
-                    }
-
-                    // Unbound players
-                    val unboundChildren = item.groupChildren.filter { !it.isBound }
-                    items(unboundChildren, key = { it.id }) { child ->
-                        GroupPlayerItem(
-                            playerId = child.id,
-                            playerName = child.name,
-                            volume = child.volume,
-                            isMuted = child.isMuted,
-                            simplePlayerAction = simplePlayerAction,
-                            bindItem = child,
-                        )
-                    }
-                }
-            }
+            GroupSettings(item, onDismiss, simplePlayerAction)
         },
         confirmButton = {}
     )
-}
-
-/**
- * Group player item with name and volume
- */
-@Composable
-private fun GroupPlayerItem(
-    playerId: String,
-    playerName: String,
-    isGroup: Boolean = false,
-    volume: Float?,
-    isMuted: Boolean?,
-    simplePlayerAction: (String, PlayerAction) -> Unit,
-    bindItem: PlayerData.Bind? = null,
-) {
-    var currentVolume by remember(volume) {
-        mutableStateOf(volume ?: 0f)
-    }
-
-    Column(
-        modifier = Modifier.fillMaxWidth(),
-        verticalArrangement = Arrangement.spacedBy((-4).dp)
-    ) {
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.SpaceBetween
-        ) {
-            Text(
-                modifier = Modifier.alpha(if (bindItem?.isBound != false) 1f else 0.4f).weight(1f),
-                text = playerName,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-                style = MaterialTheme.typography.bodyMedium,
-                fontWeight = FontWeight.Medium,
-                color = MaterialTheme.colorScheme.onSurface
-            )
-
-            // Show button only for non-current players (when bindItem is provided)
-            bindItem?.let { bind ->
-                val itemId = listOf(playerId)
-                IconButton(
-                    enabled = bindItem.isManageable,
-                    onClick = {
-                        simplePlayerAction(
-                            bind.parentId,
-                            PlayerAction.GroupManage(
-                                toAdd = itemId.takeIf { !bind.isBound },
-                                toRemove = itemId.takeIf { bind.isBound }
-                            )
-                        )
-                    }
-                ) {
-                    Icon(
-                        modifier = Modifier.alpha(if(bindItem.isManageable) 1f else 0.4f),
-                        imageVector = if (bindItem.isBound) Icons.Default.Remove else Icons.Default.Add,
-                        contentDescription = if (bindItem.isBound) "Remove from group" else "Add to group",
-                        tint = if (bindItem.isBound)
-                            MaterialTheme.colorScheme.error
-                        else
-                            MaterialTheme.colorScheme.primary
-                    )
-                }
-            }
-        }
-
-        val volumeEnabled = volume != null && bindItem?.isBound != false
-        Row {
-            isMuted?.let {
-                IconButton(onClick = {
-                    simplePlayerAction(
-                        playerId,
-                        PlayerAction.ToggleMute(isMuted)
-                    )
-                }, enabled = volumeEnabled) {
-                    Icon(
-                        imageVector = if (isMuted) Icons.AutoMirrored.Filled.VolumeMute else Icons.AutoMirrored.Filled.VolumeUp,
-                        contentDescription = if (isMuted) "Unmute" else "Mute"
-                    )
-                }
-            }
-            Slider(
-                modifier = Modifier.fillMaxWidth().alpha(if (volumeEnabled) 1f else 0.4f),
-                value = currentVolume,
-                valueRange = 0f..100f,
-                enabled = volumeEnabled,
-                onValueChange = {
-                    currentVolume = it
-                },
-                onValueChangeFinished = {
-                    simplePlayerAction(
-                        playerId,
-                        if (isGroup) PlayerAction.GroupVolumeSet(currentVolume.toDouble())
-                        else PlayerAction.VolumeSet(currentVolume.toDouble())
-                    )
-                },
-                thumb = {
-                    SliderDefaults.Thumb(
-                        interactionSource = remember { MutableInteractionSource() },
-                        thumbSize = DpSize(16.dp, 16.dp),
-                        colors = SliderDefaults.colors()
-                            .copy(thumbColor = MaterialTheme.colorScheme.secondary),
-                    )
-                },
-                track = { sliderState ->
-                    SliderDefaults.Track(
-                        sliderState = sliderState,
-                        thumbTrackGapSize = 0.dp,
-                        trackInsideCornerSize = 0.dp,
-                        drawStopIndicator = null,
-                        modifier = Modifier.height(4.dp)
-                    )
-                }
-            )
-        }
-    }
 }
 
 @Composable
